# Add lovasz hinge loss to segmentation pipeline