@glenn-jocher 2025-06-10
Thank you for the detailed experimental comparison @genji970! Your results show interesting insights - the Lovasz loss appears to provide slightly better final mAP scores (0.507 vs 0.509 at epoch 5) while using more GPU memory, which is expected behavior.

Your PR looks well-implemented and you've completed the key requirements (CLA signed, experimental validation provided). The main things to focus on now are ensuring all CI checks pass and being responsive to any feedback from the engineering team during code review. The comparative results you've shared demonstrate the feature works as intended with reasonable computational overhead.

The team will review the implementation details and may have suggestions for optimization or additional test cases, but you've done excellent groundwork with the experimental validation.