@genji970 2025-06-10
Thank you for your response. @glenn-jocher. I also think there need more additional process to add lovasz since it is hard to differentiate perfectly with coco128 etc. I will answer immediately if there are any issue!