@glenn-jocher 2025-06-11
@genji970 you're absolutely right that COCO128 may not fully showcase the benefits of Lovasz loss - it typically shows more significant improvements on datasets with challenging boundary cases or class imbalance. Your proactive attitude and willingness to iterate is exactly what we look for in contributors!

The next step is waiting for the automated CI checks to complete, then the engineering team will do a detailed code review focusing on implementation quality, edge cases, and integration with the existing codebase. Keep an eye on the PR for any test failures or review comments.